Refactor interactor factories
