+++ src/implementation/factories/interactors/location/CreateLocationFactory.php
@@ -1,22 +1,23 @@
 <?php
+
 declare(strict_types=1);
 
-namespace DrlArchive\implementation\factories\interactors;
+namespace DrlArchive\implementation\factories\interactors\location;
 
 
 use DrlArchive\core\classes\Request;
 use DrlArchive\core\factories\interactors\InteractorFactoryInterface;
-use DrlArchive\core\interactors\team\CreateTeam\CreateTeam;
+use DrlArchive\core\interactors\location\createLocation\CreateLocation;
 use DrlArchive\core\interfaces\boundaries\InteractorInterface;
 use DrlArchive\core\interfaces\boundaries\PresenterInterface;
 use DrlArchive\core\interfaces\repositories\UserRepositoryInterface;
 use DrlArchive\implementation\factories\managers\TransactionManagerFactory;
 use DrlArchive\implementation\factories\repositories\DeaneryRepositoryFactory;
+use DrlArchive\implementation\factories\repositories\LocationRepositoryFactory;
 use DrlArchive\implementation\factories\repositories\SecurityRepositoryFactory;
-use DrlArchive\implementation\factories\repositories\TeamRepositoryFactory;
 use DrlArchive\implementation\factories\repositories\UserRepositoryFactory;
 
-class CreateTeamFactory implements InteractorFactoryInterface
+class CreateLocationFactory implements InteractorFactoryInterface
 {
 
     public function create(
@@ -24,18 +25,20 @@ class CreateTeamFactory implements InteractorFactoryInterface
         ?Request $request = null,
         int $loggedInUserId = UserRepositoryInterface::GUEST_USER
     ): InteractorInterface {
-        $useCase = new CreateTeam();
+        $useCase = new CreateLocation();
         $useCase->setRequest($request);
         $useCase->setPresenter($presenter);
         $useCase->setUserRepository(
-            (new UserRepositoryFactory())->create()
+            (new UserRepositoryFactory())->create($loggedInUserId)
         );
         $useCase->setSecurityRepository(
             (new SecurityRepositoryFactory())->create()
         );
-        $useCase->setTeamRepository((new TeamRepositoryFactory())->create());
         $useCase->setDeaneryRepository(
             (new DeaneryRepositoryFactory())->create()
+        );
+        $useCase->setLocationRepository(
+            (new LocationRepositoryFactory())->create()
         );
         $useCase->setTransactionManager(
             (new TransactionManagerFactory())->create()
