+++ src/implementation/factories/interactors/location/LocationFuzzySearchFactory.php
@@ -1,0 +1,42 @@
+<?php
+
+declare(strict_types=1);
+
+namespace DrlArchive\implementation\factories\interactors\location;
+
+
+use DrlArchive\core\classes\Request;
+use DrlArchive\core\factories\interactors\InteractorFactoryInterface;
+use DrlArchive\core\interactors\location\locationFuzzySearch\LocationFuzzySearch;
+use DrlArchive\core\interfaces\boundaries\InteractorInterface;
+use DrlArchive\core\interfaces\boundaries\PresenterInterface;
+use DrlArchive\core\interfaces\repositories\UserRepositoryInterface;
+use DrlArchive\implementation\factories\repositories\LocationRepositoryFactory;
+use DrlArchive\implementation\factories\repositories\SecurityRepositoryFactory;
+use DrlArchive\implementation\factories\repositories\UserRepositoryFactory;
+
+class LocationFuzzySearchFactory implements InteractorFactoryInterface
+{
+
+    public function create(
+        PresenterInterface $presenter,
+        ?Request $request = null,
+        int $loggedInUserId = UserRepositoryInterface::GUEST_USER
+    ): InteractorInterface {
+        $useCase = new LocationFuzzySearch();
+
+        $useCase->setRequest($request);
+        $useCase->setPresenter($presenter);
+        $useCase->setLocationRepository(
+            (new LocationRepositoryFactory())->create()
+        );
+        $useCase->setUserRepository(
+            (new UserRepositoryFactory())->create($loggedInUserId)
+        );
+        $useCase->setSecurityRepository(
+            (new SecurityRepositoryFactory())->create()
+        );
+
+        return $useCase;
+    }
+}
