+++ src/implementation/factories/interactors/team/CreateTeamFactory.php
@@ -1,21 +1,22 @@
 <?php
 declare(strict_types=1);
 
-namespace DrlArchive\implementation\factories\interactors;
+namespace DrlArchive\implementation\factories\interactors\team;
 
 
 use DrlArchive\core\classes\Request;
 use DrlArchive\core\factories\interactors\InteractorFactoryInterface;
-use DrlArchive\core\interactors\competition\createDrlCompetition\CreateDrlCompetition;
+use DrlArchive\core\interactors\team\CreateTeam\CreateTeam;
 use DrlArchive\core\interfaces\boundaries\InteractorInterface;
 use DrlArchive\core\interfaces\boundaries\PresenterInterface;
 use DrlArchive\core\interfaces\repositories\UserRepositoryInterface;
 use DrlArchive\implementation\factories\managers\TransactionManagerFactory;
-use DrlArchive\implementation\factories\repositories\CompetitionRepositoryFactory;
+use DrlArchive\implementation\factories\repositories\DeaneryRepositoryFactory;
 use DrlArchive\implementation\factories\repositories\SecurityRepositoryFactory;
+use DrlArchive\implementation\factories\repositories\TeamRepositoryFactory;
 use DrlArchive\implementation\factories\repositories\UserRepositoryFactory;
 
-class CreateDrlCompetitionFactory implements InteractorFactoryInterface
+class CreateTeamFactory implements InteractorFactoryInterface
 {
 
     public function create(
@@ -23,21 +24,23 @@ class CreateDrlCompetitionFactory implements InteractorFactoryInterface
         ?Request $request = null,
         int $loggedInUserId = UserRepositoryInterface::GUEST_USER
     ): InteractorInterface {
-        $useCase = new CreateDrlCompetition();
+        $useCase = new CreateTeam();
         $useCase->setRequest($request);
         $useCase->setPresenter($presenter);
         $useCase->setUserRepository(
-            (new UserRepositoryFactory())->create($loggedInUserId)
+            (new UserRepositoryFactory())->create()
         );
         $useCase->setSecurityRepository(
             (new SecurityRepositoryFactory())->create()
         );
-        $useCase->setCompetitionRepository(
-            (new CompetitionRepositoryFactory())->create()
+        $useCase->setTeamRepository((new TeamRepositoryFactory())->create());
+        $useCase->setDeaneryRepository(
+            (new DeaneryRepositoryFactory())->create()
         );
         $useCase->setTransactionManager(
             (new TransactionManagerFactory())->create()
         );
+
         return $useCase;
     }
 }
